Add test suite to validate YASS's correctness
This commit adds a black-box test to verify that YASS is correct.

The test amounts to (i) using YASS to compute the set of models of each instance in a comprehensive set of instances, and (ii) checking that the sets of models computed by YASS are the same as the ones that were precomputed using MiniSAT and Glucose.