Reduce GC overhead using a pool for clause slices
This commit introduces the use of `sync.Pool` to manage the clauses' underlying slices. The idea is to allow new clauses to re-use the slice of a previously deleted clause without having to burden the system to allocate a new clause.

This change results in more than 10X decrease in memory need to solve some large instances. Though, the value of `nPools` still has to be optimized.