Remove allocation pool and free "dead" slices
Currently, clauses keep being referenced even after deletion (e.g. in the underlying array of the learnt slice or the the watcher slices). Because of that, many slices of literals were not properly freed by the garbage collector. This commits fixes this issue by "cutting" the reference to the slice when a clause is deleted.

As expected, this change led to a substantial reduction of in-use memory. What was not expected is that the pool of slice is now adding more overhead than help. I think the solver will still benefit of a slice freelist at some point. Though, this would likely have to be a little more tailored than sync.Pool.

Another change introduced in this commit is the use of a bit mask instead of different booleans to maintain the different status of a clause. This currently saves one byte per clause.