Speed-up propagation by tracking last watched lit
Rather than starting the search for a new literal to watch from the beginning, start the search from the position of the last previously watched literal.

The change yields a ~20% speed improvement on large instance with large clauses. Though, it's important to note that it has drawbacks. First, it introduces some overhead for small clauses (e.g. 3~6 literals). Second, it modifies the behavior of the solver as it impacts propagation order and thus what conflict are analyzed to produce new clauses.

Having different implementations for clauses of different sizes might help mitigate the first drawback.