Make prints more informative
The main change in this commit is the addition of an "event" char at the beginning of each row. The letter indicates if the print was a restart (R), a clean-up of the local clauses (C), or simple info (no letter). This removed the need for the restart column which has been replaced with metrics about the core clauses.